RAG/requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.13
python-dotenv==1.0.0
chromadb==0.4.22
faiss-cpu==1.7.4
tiktoken==0.5.2
pypdf==3.17.1
sentence-transformers==2.2.2
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6 